Added Notifications & Labels
Used xPlayer.showNotification() to draw a notification every time a player receives a reward. I have also added labels into the configuration area so that notifications do not show spawn codes which are not grammatically correct and look bad. 

Next update: Adding a KeyMapping for 'Z' to run the ShowXP command, just like in GTA:O where you can press Z to see your XP and the Leaderboard. I'll also be adding a configuration as well as the function for a chat message which will alert everyone in the server that a player has reached the next rank. 

Another thing I want to implement soon which will be configurable is giving a player a default reward if their rank isn't in the Rewards Table.

Coming Soon: Using ESX Vehicle Shop to allow vehicles to be given as rewards.